server timeouts: improved dead timeout handling again
The previous behavior based on timestamps of successful write() and read()
calls did not really work, as it's quite possible that write() will
successfully send data somewhere even if the remote end never handles it and
responds to it.  Now dead timeout is handled differently: dead timeout is
set when the first write is attempted and updated when data is read from the
server until the last pending response has been received at which point it's
reset back to 0.